refactor: 아토믹디자인으로 변경
