--- packages/mobile/src/components/Home/style.module.scss
@@ -1,92 +1,0 @@
-.home{
-    header{
-        display: flex;
-        padding:48px 0 43px 0;
-        justify-content: space-between;
-        align-items: center;
-        h1{
-            font-weight: 500;
-            font-size: 23px;
-            line-height: 33px;
-        }
-        span{
-            color:#9d9d9d;
-        }
-    }
-    .schedule{
-        display: flex;
-        flex-wrap: nowrap; 
-        overflow-x: auto;
-        div{
-            display:flex;
-            align-items: center;
-            justify-content: space-between;
-            padding-left:20px;
-            padding-right:20px;
-            margin-right: 23px;
-            min-width:70%;
-    
-        }
-        p{
-            line-height: 24px;
-            font-weight: 500;
-            padding-right:10px;
-        }
-    }
-    .external-information{
-        display: flex;
-        margin-top:31px;
-        font-size:11px;
-        div{
-            display: flex;
-            flex-direction: column;
-            align-items: center;
-            justify-content: space-evenly;
-            width:100%;
-            
-        }
-        div:first-child{
-            margin-right: 25px;
-        }
-.more{
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    svg{
-        margin-right:4px;
-    }
-
-    svg path{
-        stroke: #222;
-    }
-
-}        
-
-    }
-    .cafeteria{
-      width:100%;  
-      margin-top:31px;
-      
-      >div{
-        padding:17px;
-      }
-      .title{
-          display: flex;
-          font-weight: 500;
-          p{
-              margin-right:5px;
-
-          }
-      }
-      .content{
-        margin-top:21px;
-          font-size: 11px;
-          line-height: 25px;
-      }
-
-    }
-
-    .notification{
-        margin-top:31px;
-    }
-}
