+++ packages/mobile/src/components/atoms/BorderBox/style.module.scss
@@ -1,0 +1,4 @@
+.border-box {
+    background: #fff;
+    border-radius: 18px;
+}
