+++ packages/mobile/src/components/atoms/Line/style.module.scss
@@ -1,0 +1,5 @@
+.line {
+    border-top: 1px solid #b4b4b4;
+    width: 100%;
+    height: 1px;
+}
