+++ packages/mobile/src/components/atoms/NavigationLink/style.module.scss
@@ -1,0 +1,25 @@
+svg path {
+    stroke: #d6d6d6;
+}
+
+.active {
+    svg path {
+        stroke: #ce5455;
+    }
+    p {
+        color: #ce5455;
+    }
+}
+
+a {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+
+    p {
+        margin-top: 6px;
+        font-size: 9px;
+        color: #d6d6d6;
+    }
+}
