+++ packages/mobile/src/components/atoms/icon/Covid.tsx
