+++ packages/mobile/src/components/atoms/icon/Food.tsx
