+++ packages/mobile/src/components/atoms/icon/Home.tsx
