+++ packages/mobile/src/components/atoms/icon/Info.tsx
