+++ packages/mobile/src/components/atoms/icon/Setting.tsx
