+++ packages/mobile/src/components/atoms/icon/Speaker.tsx
