+++ packages/mobile/src/components/atoms/icon/Write.tsx
