+++ packages/mobile/src/components/atoms/icon/가끔눈_밤.tsx
