+++ packages/mobile/src/components/atoms/icon/가끔눈또는비_밤.tsx
