+++ packages/mobile/src/components/atoms/icon/가끔비_낮.tsx
