+++ packages/mobile/src/components/atoms/icon/가끔비_밤.tsx
