+++ packages/mobile/src/components/atoms/icon/가끔비또는눈.tsx
