+++ packages/mobile/src/components/atoms/icon/구름많음_낮.tsx
