+++ packages/mobile/src/components/atoms/icon/구름많음_밤.tsx
