+++ packages/mobile/src/components/atoms/icon/구름조금_낮.tsx
