+++ packages/mobile/src/components/atoms/icon/눈또는비.tsx
