+++ packages/mobile/src/components/atoms/icon/맑음_낮.tsx
