+++ packages/mobile/src/components/atoms/icon/맑음_밤.tsx
