+++ packages/mobile/src/components/atoms/icon/박무.tsx
