+++ packages/mobile/src/components/atoms/icon/비.tsx
