+++ packages/mobile/src/components/atoms/icon/비또는눈.tsx
