+++ packages/mobile/src/components/atoms/icon/소나기.tsx
