+++ packages/mobile/src/components/atoms/icon/안개.tsx
