+++ packages/mobile/src/components/atoms/icon/연무.tsx
