+++ packages/mobile/src/components/atoms/icon/천둥번개.tsx
