+++ packages/mobile/src/components/atoms/icon/황사.tsx
