+++ packages/mobile/src/components/atoms/icon/흐림.tsx
