+++ packages/mobile/src/components/molecules/Footer/style.module.scss
@@ -1,0 +1,13 @@
+.footer {
+    position: fixed;
+    bottom: 0;
+    left: 0;
+    right: 0;
+    width: 100%;
+    height: 58px;
+    display: flex;
+    justify-content: space-around;
+    background: #fff;
+    box-shadow: 0px -1px 10px rgba(0, 0, 0, 0.03);
+    border-radius: 15px 15px 0px 0px;
+}
