--- packages/mobile/src/components/shared/BorderBox/style.module.scss
@@ -1,4 +1,0 @@
-.border-box{
-    background:#fff;
-    border-radius:18px;
-}
