--- packages/mobile/src/components/shared/DefaultPageTemplate/style.module.scss
@@ -1,5 +1,0 @@
-.default-page-template{
-    padding-left:25px;
-    padding-right:25px;
-    overflow-y: auto;
-}
