--- packages/mobile/src/components/shared/Footer/style.module.scss
@@ -1,40 +1,0 @@
-
-.footer{
-    position:fixed;
-    bottom:0;
-    left:0;
-    right:0;
-    width:100%;
-    height:58px;
-
-    display:flex;
-    justify-content: space-around;
-    background: #fff;
-    box-shadow: 0px -1px 10px rgba(0, 0, 0, 0.03);
-    border-radius: 15px 15px 0px 0px;
-
-    svg path{
-        stroke:#D6D6D6;
-    }
-
-    .active{
-        svg path{
-            stroke:#ce5455;
-        }
-        p{
-            color:#ce5455;
-        }
-    }
-    a{
-        display: flex;
-        flex-direction: column;
-        justify-content: center;
-        align-items: center;
-
-        p{
-            margin-top:6px;
-            font-size: 9px;
-            color:#d6d6d6;
-        }
-    }
-}
