+++ packages/mobile/src/components/templates/DefaultPageTemplate/index.tsx
