+++ packages/mobile/src/components/templates/DefaultPageTemplate/style.module.scss
@@ -1,0 +1,6 @@
+.default-page-template {
+    padding-left: 25px;
+    padding-right: 25px;
+    padding-bottom: 98px;
+    overflow-y: auto;
+}
