+++ packages/mobile/src/page/Home/style.module.scss
@@ -1,0 +1,89 @@
+.home {
+    header {
+        display: flex;
+        padding: 48px 0 43px 0;
+        justify-content: space-between;
+        align-items: center;
+        h1 {
+            font-weight: 500;
+            font-size: 23px;
+            line-height: 33px;
+        }
+    }
+    .schedule {
+        display: flex;
+        flex-wrap: nowrap;
+        overflow-x: auto;
+        div {
+            display: flex;
+            align-items: center;
+            justify-content: space-between;
+            padding-left: 20px;
+            padding-right: 20px;
+            margin-right: 23px;
+            min-width: 70%;
+        }
+        p {
+            line-height: 24px;
+            font-weight: 500;
+            padding-right: 10px;
+        }
+    }
+    .external-information {
+        display: flex;
+        margin-top: 31px;
+        font-size: 11px;
+        div {
+            display: flex;
+            flex-direction: column;
+            align-items: center;
+            justify-content: space-evenly;
+            width: 100%;
+        }
+        div:first-child {
+            margin-right: 25px;
+        }
+        .more {
+            display: flex;
+            flex-direction: row;
+            justify-content: center;
+            svg {
+                margin-right: 4px;
+            }
+
+            svg path {
+                stroke: #222;
+            }
+        }
+        .important{
+            font-size: 18px;
+            font-weight: 500;
+        }
+    }
+    .cafeteria {
+        width: 100%;
+        margin-top: 31px;
+
+        > div {
+            padding: 17px;
+        }
+        .title {
+            display: flex;
+            font-weight: 500;
+            margin-bottom: 16px;
+
+            p {
+                margin-right: 5px;
+            }
+        }
+        .content {
+            padding: 21px 20px 0px 0px;
+            font-size: 11px;
+            line-height: 25px;
+        }
+    }
+
+    .notification {
+        margin-top: 31px;
+    }
+}
